learn-pr/student-evangelism/python-flask-build-ai-web-app/source-final/requirements.txt
requests
Flask
azure-cognitiveservices-vision-computervision
